kafka_app/requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
idna==3.1
kafka-python==2.0.2
multidict==5.1.0
pydantic==1.8.1
typing-extensions==3.7.4.3
yarl==1.6.3
